% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.int.effects.R
\name{estim.int.effects}
\alias{estim.int.effects}
\title{Estimate mean results, relative risks and differences}
\usage{
estim.int.effects(
  ltmle_MSM = ltmle_MSM,
  estimator = c("gcomp", "iptw", "tmle")
)
}
\arguments{
\item{ltmle_MSM}{an output from \code{int.ltmleMSM} function}

\item{estimator}{estimator of the marginal interaction effect. One of "gcomp", "iptw" or "tmle".}
}
\description{
Estimate mean results, relative risks and differences
}
