% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.param.R
\name{generate.data}
\alias{generate.data}
\title{Simulate an illustrative data set}
\usage{
generate.data(N, b = param.causal.model())
}
\arguments{
\item{N}{Sample size of the illustrative data set.}

\item{b}{A list of 4 vectors of parameters,}
}
\value{
A data frame of size \code{N} using the parameters \code{b},
generated from the following structural causal model:

\eqn{P(L_1 = 1) = p_{L_1}}

\eqn{P(L_2 = 1) = p_{L_2}}

\eqn{P(L_3 = 1) = p_{L_3}}

\eqn{P(A_1 = 1) = \beta_{A_1} + \beta_{L_1,A_1} L_1 + \beta_{L_2,A_1} L_2}

\eqn{P(A_2 = 1) = \beta_{A_2} + \beta_{L_1,A_2} L_1 + \beta_{L_3,A_2} L_3}

\eqn{P(Y = 1) = \beta_{Y} + \beta_{L_1,Y} L_1 + \beta_{L_2,Y} L_2 + \beta_{L_3,Y} L_3 + \beta_{A_1,Y} A_1 + \beta_{A_2,Y} A_2 + \beta_{A_1 \ast A_2,Y} (A_1 \ast A_2)}

The names of the baseline confounders \eqn{L_1, L_2, L_3} are \code{conf1}, \code{conf2}
and \code{conf3}. The names of the exposure \eqn{A_1} is \code{sex}, the name of
the exposure \eqn{A_2} is \code{env}. The name of the outcome \eqn{Y} is
\code{hlth.outcome}.
}
\description{
Simulate an illustrative data set of size \code{N} using the parameters \code{b}
defined using the \code{param.causal.model()} function.
}
\examples{
set.seed(12345)
df <- generate.data(N = 1000, b = param.causal.model())
head(df)
}
