% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.ltmleMSM.R
\name{int.ltmleMSM}
\alias{int.ltmleMSM}
\title{Fitting a Marginal Structural Model to estimate marginal interaction effects}
\usage{
int.ltmleMSM(
  data = data,
  Anodes = Anodes,
  Cnodes = NULL,
  Lnodes = NULL,
  Ynodes = Ynodes,
  survivalOutcome = FALSE,
  Qform = Qform,
  gform = gform,
  gbounds = c(0.01, 1),
  Yrange = NULL,
  deterministic.g.function = NULL,
  SL.library = list(Q = "SL.glm", g = "SL.glm"),
  SL.cvControl = list(),
  final.Ynodes = NULL,
  stratify = FALSE,
  msm.weights = "empirical",
  estimate.time = FALSE,
  gcomp = gcomp,
  iptw.only = iptw.only,
  deterministic.Q.function = NULL,
  variance.method = "ic",
  observation.weights = NULL,
  id = NULL,
  B = 2000,
  boot.seed = NULL
)
}
\arguments{
\item{data}{data frame following the time-ordering of the nodes. Should follow the format recommended for the ltmle package}

\item{Anodes}{column names in \code{data} for the two exposures \code{c(A1,A2)}}

\item{Cnodes}{used for censoring nodes in ltmleMSM function. \code{NULL} by default, survival is not yet implemented for the \code{int.ltmleMSM function}}

\item{Lnodes}{column names in \code{data} for confounders of the A1 -> Y and A2 -> Y relationships}

\item{Ynodes}{column names in \code{data} for the outcome node}

\item{survivalOutcome}{column names in \code{data} for the outcome node. \code{FALSE} by default, survival is not yet implemented for the \code{int.ltmleMSM function}}

\item{boot.seed}{}
}
\value{
\code{int.ltmleMSM} returns 3 objects:
\itemize{ \item \code{ltmle_MSM} the output from the ltmleMSM function
\item \code{df.int} the data frame where the exposures names are \code{c(A1,A2)} and the outcome name is \code{Y}}
\item \code{bootstrap.res} a data frame containing estimation from bootstrap samples (for g-computation only)
}
\description{
\code{int.ltmleMSM} is used to fit a MSM using the ltmleMSM function, from the \code{ltmle} package.
}
\details{
Details to detail..
}
\examples{
set.seed(12345)
b = param.causal.model()
df <- generate.data(N = 10000, b = b)
summary(df)

# Define Q and g formulas
# an A1 * A2 interaction term is recommended in the Q formula for the estimation of interaction effects
Q_formulas = c(Y="Q.kplus1 ~ conf.fac1 + conf.fac2 + conf.fac3 + A1 * A2")
g_formulas = c("A1 ~ conf.fac1 + conf.fac2",
               "A2 ~ conf.fac1 + conf.fac3")

# Define SuperLearner libraries
SL.library = list(Q=list("SL.glm", c("SL.glm", "screen.corP"),"SL.glmnet", "SL.mean"),
                  g=list("SL.glm", c("SL.glm", "screen.corP"),"SL.glmnet", "SL.mean"))

# Estimate MSM parameters by IPTW and TMLE
interaction.ltmle <- int.ltmleMSM(data = df,
                                  Qform = Q_formulas,
                                  gform = g_formulas,
                                  Anodes = c("sex", "env"),
                                  Lnodes = c("conf.fac1", "conf.fac2", "conf.fac3"),
                                  Ynodes = c("hlth.outcome"),
                                  SL.library = SL.library,
                                  gcomp = FALSE,
                                  iptw.only = FALSE,
                                  survivalOutcome = FALSE,
                                  variance.method = "ic")

# Estimate MSM parameters by g-computation
interaction.gcomp <- int.ltmleMSM(data = df,
                                  Qform = Q_formulas,
                                  gform = g_formulas,
                                  Anodes = c("sex", "env"),
                                  Lnodes = c("conf.fac1", "conf.fac2", "conf.fac3"),
                                  Ynodes = c("hlth.outcome"),
                                  SL.library = list(Q="SL.lm", g="SL.mean"),
                                  gcomp = TRUE,
                                  iptw.only = FALSE,
                                  survivalOutcome = FALSE,
                                  variance.method = "ic",
                                  B = 100,
                                  boot.seed = 54321)
}
