% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MargIntTmle-package.R
\docType{package}
\name{MargIntTmle-package}
\alias{MargIntTmle}
\alias{MargIntTmle-package}
\title{MargIntTmle: Marginal interaction estimation using g-computation, IPTW or TMLE}
\description{
Functions to estimate the marginal interaction effect of 2 categorical variables on an outcome, and to describe the results according to Knol and VanderWeele recommendations (2012).
}
\author{
\strong{Maintainer}: Benoit Lepage \email{benoit.lepage@univ-tlse3.fr} (\href{https://orcid.org/0000-0002-1586-2055}{ORCID})

}
\keyword{internal}
