% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.param.R
\name{param.causal.model}
\alias{param.causal.model}
\title{Define parameters to simulate data}
\usage{
param.causal.model(
  p_L1 = 0.5,
  p_L2 = 0.2,
  p_L3 = 0.7,
  b_A1 = 0.1,
  b_L1_A1 = 0.15,
  b_L2_A1 = 0.25,
  b_A2 = 0.15,
  b_L1_A2 = 0.2,
  b_L3_A2 = 0.2,
  Y_type = "binary",
  b_Y = 0.1,
  b_L1_Y = 0.02,
  b_L2_Y = 0.02,
  b_L3_Y = -0.02,
  b_A1_Y = 0.3,
  b_A2_Y = 0.1,
  b_A1A2_Y = 0.4,
  se_Y = NULL
)
}
\arguments{
\item{p_L1}{Probability \eqn{P(L_1 = 1) = p_{L_1}}.}

\item{p_L2}{Probability \eqn{P(L_2 = 1) = p_{L_2}}.}

\item{p_L3}{Probability \eqn{P(L_3 = 1) = p_{L_3}}.}

\item{b_A1}{Probability \eqn{P(A_1 = 1|L_1 = 0, L_2 = 0) = \beta_{A_1}}.}

\item{b_L1_A1}{Additive effect \eqn{\beta_{L_1,A_1}} of \eqn{L_1 \rightarrow A_1}.}

\item{b_L2_A1}{Additive effect \eqn{\beta_{L_2,A_1}} of \eqn{L_2 \rightarrow A_1}.}

\item{b_A2}{Probability \eqn{P(A_2 = 1|L_1 = 0, L_3 = 0) = \beta_{A_2}}.}

\item{b_L1_A2}{Additive effect \eqn{\beta_{L_1,A_2}} of \eqn{L_1 \rightarrow A_2}.}

\item{b_L3_A2}{Additive effect \eqn{\beta_{L_3,A_2}} of \eqn{L_3 \rightarrow A_2}.}

\item{Y_type}{Is Y "binary" or "continuous"?}

\item{b_Y}{Probability \eqn{P(Y = 1|L_1 = 0, L_2 = 0, L_3 = 0, A_1 = 0, A_2 = 0) = \beta_{Y}} if Y is binary;
or expectation \eqn{\mathbb{E}(Y|L_1 = 0, L_2 = 0, L_3 = 0, A_1 = 0, A_2 = 0) = \beta_{Y}}}

\item{b_L1_Y}{Additive effect \eqn{\beta_{L_1,Y}} of \eqn{L_1 \rightarrow Y}.}

\item{b_L2_Y}{Additive effect \eqn{\beta_{L_2,Y}} of \eqn{L_2 \rightarrow Y}.}

\item{b_L3_Y}{Additive effect \eqn{\beta_{L_3,Y}} of \eqn{L_3 \rightarrow Y}.}

\item{b_A1_Y}{Additive effect \eqn{\beta_{A_1,Y}} of \eqn{A_1 \rightarrow Y}.}

\item{b_A2_Y}{Additive effect \eqn{\beta_{A_2,Y}} of \eqn{A_2 \rightarrow Y}.}

\item{b_A1A2_Y}{Interaction additive effect \eqn{\beta_{A_1 \ast A_2,Y}} of \eqn{(A_1 \ast A_2) \rightarrow Y}.}

\item{se_Y}{Standard error of the outcome for quantitative outcomes, should be null for binary outcomes.}
}
\value{
A list of 4 vectors of parameters.
}
\description{
This function defines parameters that are used to simulate illustrative data sets
consisting of three baseline confounders \eqn{L_1, L_2, L_3}, two exposures of interest
\eqn{A_1, A_2} and one outcome \eqn{Y}. All the variables are binary.
}
\examples{
param.causal.model()
}
