% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.int.R
\name{summary.int}
\alias{summary.int}
\title{Estimate mean results, relative risks and differences}
\usage{
\method{summary}{int}(ltmle_MSM = ltmle_MSM, estimator = c("gcomp", "iptw", "tmle"))
}
\arguments{
\item{estimator}{}
}
\description{
Estimate mean results, relative risks and differences
}
