% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.int.R
\name{out.int.fig}
\alias{out.int.fig}
\title{Create plot of the marginal interaction effects}
\usage{
out.int.fig(int.r = int.r)
}
\arguments{
\item{int.r}{}
}
\description{
Create plot of the marginal interaction effects
}
