% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.int.R
\name{out.int.table}
\alias{out.int.table}
\title{Create table of the marginal interaction effects}
\usage{
out.int.table(int.r = int.r)
}
\arguments{
\item{int.r}{}
}
\description{
Create table of the marginal interaction effects
}
