% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.int.R
\name{summary.int}
\alias{summary.int}
\title{Estimate mean results, relative risks and differences}
\usage{
\method{summary}{int}(ltmle_MSM = ltmle_MSM, estimator = c("gcomp", "iptw", "tmle"))
}
\arguments{
\item{ltmle_MSM}{an output from \code{int.ltmleMSM} function}

\item{estimator}{estimator of the marginal interaction effect. One of "gcomp", "iptw" or "tmle".}
}
\description{
Estimate mean results, relative risks and differences
}
